Don't use vendor/ in tests
To work around the fact that Ginkgo's gexec.Build() does not yet know
about vgo modules. This means my tests choke on unexpected output
coming from module downloads.

Switching to -mod=readonly is slower and (I think) less trustworthy
because it needs to make a network call to download all the dependencies
every time. But tests are better than not-tests.

[jchesterpivotal/concourse-build-resource#8]

Signed-off-by: Jacques Chester <jchester@pivotal.io>