Switch existing K-V files to use underscores in names
Previously, I was using the hyphen. This is fine, but for consistency
it should be one or the other. Since I've used underscores in the YAML,
and since underscores tend to be less problematic in programming
language contexts (not mistaken for minus operation), I've switched the
existing files (eg. job-number) over.

[jchesterpivotal/concourse-build-resource#17]

Signed-off-by: Jacques Chester <jchester@pivotal.io>