Create 'when' for fully-specified in case
This is the original "happy path" for the resource: all of the
source information (concourse url, team, pipeline, job) are
specified. But in jchesterpivotal/concourse-build-resource#6, it
became possible to selectively specify only some of this information.

But at the time I did not add tests to deal with missing values.
The jobs endpoint can't be accessed without information about the
pipeline name and job name. If these are not provided by the source
struct, we will need to extract it from the build struct before
proceeding, or alternatively simply skip writing the file (as we
do when the events endpoint is forbidden).

This commit simply creates a top-level 'when' block and hoists
all the existing happy-path tests into it, as a precursor to adding
more specific scenarios.

[jchesterpivotal/concourse-build-resource#26]

Signed-off-by: Jacques Chester <jacques@chester.id.au>