Use file copying instead of copied code
Each endpoint's .json or .log file is written out thrice. The first
time is the raw kind, the other two carry build number or build information.

Previously I achieved this by writing them out entirely for each variation.
In this commit it's written once and then copied. This should particularly
help with the readability of the events code.

[jchesterpivotal/concourse-build-resource#19]

Signed-off-by: Jacques Chester <jacques@chester.id.au>